% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head-models.R
\name{constrainOrientation}
\alias{constrainOrientation}
\title{Constrain a free-orientation leadfield to surface normals}
\usage{
constrainOrientation(leadfield, normals)
}
\arguments{
\item{leadfield}{An n_electrodes x (n_sources * 3) free-orientation leadfield.}

\item{normals}{An n_sources x 3 matrix of unit source normals.}
}
\value{
An n_electrodes x n_sources orientation-constrained leadfield.
}
\description{
Collapses the three orientation columns of each source to a single
surface-normal-oriented column, giving an n_electrodes x n_sources leadfield.
}
\examples{
ss <- sphericalSourceSpace(40, radius = 0.6)
e <- icosphere(2)$vertices[1:19, ]
lf <- sphereLeadfield(e, ss$positions)
lc <- constrainOrientation(lf, ss$normals)
dim(lc)
}
