% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head-models.R
\name{solidAngle}
\alias{solidAngle}
\title{Solid angle subtended by a triangle at a point (van Oosterom-Strackee)}
\usage{
solidAngle(v1, v2, v3, r)
}
\arguments{
\item{v1, v2, v3}{Numeric length-3 triangle vertices.}

\item{r}{Numeric length-3 observation point.}
}
\value{
The signed solid angle in steradians.
}
\description{
Solid angle subtended by a triangle at a point (van Oosterom-Strackee)
}
\examples{
# One face of the unit octahedron, seen from the centre, spans an eighth
# of the full sphere: pi / 2 steradians.
solidAngle(c(1, 0, 0), c(0, 1, 0), c(0, 0, 1), c(0, 0, 0))
}
